% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthapower.R
\name{mthapower}
\alias{mthapower}
\title{Post-hoc power calculations - mtDNA haplogroups}
\usage{
mthapower(n.cases = ncases, p0 = p0, Nh = Nh, OR.cas.ctrl = OR.cas.ctrl,
  sig.level = sig.level)
}
\arguments{
\item{n.cases}{number of cases or controls from the study. It can be either a single value, or a sequence: \code{n.cases <- 300; n.cases <- seq(50,500 by=10)}}

\item{p0}{the frequency of the haplogroup in the control population, the controls among exposed. It depends on haplogroup baseline frequency.}

\item{Nh}{number of categories for haplogroups. Usually 10 haplogroups plus one category for rare haplogroups: \code{Nh <- 11}.}

\item{OR.cas.ctrl}{(p1 / (1-p1)) / (p0 / (1-p0)) the OR you want to detect with your data.}

\item{sig.level}{error alpha accepted. Can take 3 possible values: 0.05, 0.01 and 0.001 (see [Table 2] of Samuels et al.)}
}
\value{
Calculates power given number of cases and other parameters. The output is an object of class data.frame, ready to plot.
}
\description{
For a given study size, determine the minimum effect size that can be detected with the desired power and significance level, in a study with \code{Nh} haplogroups.
Note: I assume that case-control equations may be valid for cohorts, equally "divided" between cases and controls.
}
\examples{
# Example 1:
pow <- mthapower(n.cases=203, p0=0.443, Nh=13, OR.cas.ctrl=2.33, sig.level=0.05)

# Example 2:
# Create data frames
pow.H150 <- mthapower(n.cases=seq(50,1000,by=50), p0=0.433, Nh=11,
                      OR.cas.ctrl=1.5, sig.level=0.05)
pow.H175 <- mthapower(n.cases=seq(50,1000,by=50), p0=0.433, Nh=11,
                      OR.cas.ctrl=1.75, sig.level=0.05)
pow.H200 <- mthapower(n.cases=seq(50,1000,by=50), p0=0.433, Nh=11,
                      OR.cas.ctrl=2, sig.level=0.05)
pow.H250 <- mthapower(n.cases=seq(50,1000,by=50), p0=0.433, Nh=11,
                      OR.cas.ctrl=2.5, sig.level=0.05)

# Bind the three data frames:
bindata <- rbind(pow.H150,pow.H175,pow.H200,pow.H250)
# Adds column OR to binded data frame:
bindata$OR <- rep(factor(c(1.50,1.75,2,2.5)),
              times = c(nrow(pow.H150),
                        nrow(pow.H175),
                        nrow(pow.H200),
                        nrow(pow.H250)))
# Create plot:
# install.packages("car")
library(car)
scatterplot(power~ncases | OR, reg.line=FALSE,
            smooth=FALSE, spread=FALSE,
            boxplots=FALSE, span=0.25, by.groups=FALSE,
            data=bindata)

}
\references{
1. David C. Samuels, Andrew D. Carothers, Robin Horton, Patrick F. Chinnery. The Power to Detect Disease Associations with Mitochondrial DNA Haplogroups. \url{http://www.cell.com/ajhg/fulltext/S0002-9297(07)63709-4}

2. Source: \url{https://github.com/aurora-mareviv/mthapower}

3. Shiny app: \url{https://aurora.shinyapps.io/mtDNA_power_calc}
}
\author{
Author and maintainer: Aurora Baluja. Email: \email{mariauror@gmail.com}
}
