% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthacases.R
\name{mthacases}
\alias{mthacases}
\title{Sample size calculations - mtDNA haplogroups}
\usage{
mthacases(p0 = p0, Nh = Nh, OR.cas.ctrl = OR.cas.ctrl, power = power,
  sig.level = sig.level)
}
\arguments{
\item{p0}{the frequency of the haplogroup in the control population, the controls among exposed. It depends on haplogroup baseline frequency.}

\item{Nh}{number of categories for haplogroups. Usually 10 haplogroups plus one category for rare haplogrs.: Nh <- 11}

\item{OR.cas.ctrl}{\code{(p1 / (1-p1)) / (p0 / (1-p0))} the OR you want to detect with your data. It can be either a single value, or a sequence: \code{OR.cas.ctrl <- 2; OR.cas.ctrl <- seq(1.25,3 by=0.5)}}

\item{power}{the power I want to detect any OR in my study.}

\item{sig.level}{error alpha accepted. Can take 3 possible values: 0.05, 0.01 and 0.001 (see [Table 2] of Samuels et al.)}

\item{cases.min}{number of cases or controls that I need to recruit.}
}
\value{
Gives the result in a data frame, easy to print in a plot.
}
\description{
Determine the minimum number of cases (\code{Ncmin}), required to detect: either a change from p0 to p1, or a given OR, with a predefined confidence interval, in a study with \code{Nh} haplogroups.
Note: I assume that case-control equations are valid for a cohort.
This function may not be generalizable for all studies.
}
\examples{
mydata <- mthacases(p0=0.445, Nh=11,
                    OR.cas.ctrl=c(2), power=80,
                    sig.level=0.05) # Baudouin study
mydata <- mthacases(p0=0.445, Nh=11,
                    OR.cas.ctrl=c(1.25,1.5,1.75,2,2.25,2.5,2.75,3),
                    power=80, sig.level=0.05)
mydata <- mydata[c(2,6)]
mydata
plot(mydata)

}
\references{
\url{http://www.cell.com/ajhg/fulltext/S0002-9297(07)63709-4}

\url{https://github.com/aurora-mareviv/mthapower}
}
\author{
Aurora Baluja, \email{mariauror@gmail.com}
}
